articular-core: reintroduced logic into UpdatableEntity for style freedom
