+++ articular-es/src/main/java/articular/core/Entity.java
@@ -32,7 +32,7 @@
 package articular.core;
 
 import articular.core.component.Component;
-import java.util.Map;
+import java.util.Collection;
 
 /**
  * Represents a game entity with some {@link Component}s.
@@ -46,49 +46,30 @@ import java.util.Map;
  * A Map of components must be provided via a DI pattern.
  * </p>
  *
- * @param <I> the input type of the game loop
+ * @param <T> the type of the game component objects
  * @author pavl_g
  */
-public interface Entity<I> extends Component<I> {
+public interface Entity<T> extends Component<T> {
 
     /**
-     * A Map of game entity components to be provided.
+     * A Collection of game entity components to be provided.
      *
-     * <p>
-     * Override and provide a map of components for this entity via the DI pattern.
-     * </p>
-     *
-     * @return a map of components by their identifiers (not null)
+     * @return a collection of game entity's components
      */
-    Map<? super Number, Component<I>> getComponents();
+    Collection<Component<T>> getComponents();
+
+    /**
+     * Retrieves a game entity's component by its identifier.
+     *
+     * @param id the component identifier
+     * @return a game entity component or a game sub-entity
+     */
+    Component<T> getComponent(Component.Id id);
 
     /**
      * Retrieves the name of this entity.
      *
-     * <p>
-     * Override and provide a name for this entity via the DI pattern.
-     * </p>
-     *
-     * @return the name of this entity (not null)
+     * @return the name of this entity (nullable)
      */
     String getName();
-
-    /**
-     * Retrieves a specific game entity component by its identifier.
-     *
-     * @param id the game entity component identifier
-     * @return a game entity component or null if it doesn't exist (nullable)
-     */
-    default Component<I> getComponent(Component.Id id) {
-        return getComponents().get(id);
-    }
-
-    /**
-     * Updates this game entity through updating its components.
-     *
-     * @param input the input to the game loop pattern
-     */
-    default void update(I input) {
-        getComponents().forEach((id, component) -> component.update(input));
-    }
 }
